moving the cleaning and restructuring to a mixin
